Fix set_params on entire subpipelines
MultiPipeline uses __getattr__ to expose subpipelines stored in the dict
MutliPipeline.pipelines as attributes, so sklearn can access them when
using .get_params and .set_params. However, __getattr__ is only called
if the attribute isn't found in the objects .__dict__ - so if a whole
pipeline is overwritten with pipeline.set_params, it silently creates a
new attribute that is accessed instead of the original subpipeline with
using getattr, but the original pipeline remains unchanged when
accessed via MutliPipeline.pipelines (which is how it's called in
.transform or .fit).

In other words, using .set_params for an entire subpipeline causes
get_params to show different parameters to what's being called in the
code.

This change adds a failing test &  a fix: implementing __setattr__ so
that the attributes and pipelines are always in sync regardless of how
they are updated.