Move version to be PEP 440 compliant
This is to support poetry: currently we're using SemVer, but installing
dev versions via poetry gives a build error, then says it's installed it
but never actually does. The PyPI released versions will be both SemVer
_and_ PEP 440 compliant. Potentially if we move some other projects away
from poetry then we could change this back.

The version can only have a major, minor and patch, an optional dev
version and a "local" version after a +.

We move the number of commits after a tag to the dev version, and the
hash & dirty flag to the local version.